requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Faker==13.3.0
idna==3.3
iniconfig==1.1.1
jsonpath-rw==1.4.0
packaging==21.3
pluggy==1.0.0
ply==3.11
py==1.11.0
PyHamcrest==2.0.3
pyparsing==3.0.7
pytest==7.0.1
pytest-env==0.6.2
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.8
